Basic structure for physics.
Currently physics is manually called, will be moved to physics manager in future once I write that
IPhysicsObject for any objects that have physics, this has basic functionality for rigid body's velocity inside it, but if a specific physics type changes how velocity is applied this can be overridden;
Added Texture component, this interfaces with sprite Renderer.
Added  bool hasComponent<TYPE>() function to gameObject, returns true if a component of the given type is present
Addded physicsUpdate(float dt); function to gameObjects this calls physics update on all IphysicsObject inherited components in itself, and all it's children.