Improves rendering and collision handling
Adds z-height sorting to the render queue, ensuring correct rendering order. elements with the same z-height will be rendered from top to bottom, if multiple children the child that was added first will be rendered first.

Modifies collision handling to use the scene root's children, removing the need for the scene's game_object list

Fixes scaling issues in the Transform component. (I was a silly when re-writing to use vector2f's...