Adds configurable game settings
Allows configuring window settings such as title, resolution, framerate, and physics timestep.

Uses prefabs in the test scene to demonstrate their usage.

This should be the final commit here for a while, as I am going to move to creating the platformer with this system, likely future commits for a while will just be bug fixes (or usability features) as I discover issues with this system.